API Reference

Elements

NNFEM.FiniteStrainContinuum — Type

FiniteStrainContinuum

Implements the finite strain element.

eledim: spatial dimension of the element (default = 2).

mat: constitutive law, a length #elem vector of materials such as PlaneStress

elnodes: the node indices in this finite element, an integer array

props: property dictionary

coords: coordinates of the vertices of the element

dhdx, weights, hs: data for integral

stress: stress at each quadrature points

Materials

NNFEM.PlaneStrain — Type

NNFEM.PlaneStressIncompressibleRivlinSaunders — Type

Pascon, João Paulo. "Large deformation analysis of plane-stress hyperelastic problems via triangular membrane finite elements." International Journal of Advanced Structural Engineering (2019): 1-20.

Assembly

NNFEM.assembleInternalForce — Function

Numerically assemble internal force vector, compute local internal force f_int 
from domain.state and then assemble to F_int
- 'globdat': GlobalData
- 'domain': Domain, finite element domain, for data structure
- 'Δt':  Float64, current time step size
Return F_int: Float64[neqs], internal force vector

NNFEM.assembleMassMatrix! — Method

compute constant sparse mass matrix
due to the time-dependent Dirichlet boundary condition
mass matrix = M,    MID
              MID'  MDD

- 'globdat': GlobalData
- 'domain': Domain, finite element domain, for data structure
here M is Float64[neqns, neqns]
     MID is Float64[neqns, nd1]
     Mlumped is Float64[neqns]

update M and MID and Mlumpe in globaldat

NNFEM.assembleStiffAndForce — Function

Numerically assemble internal force vector and stiffness matrix, compute local internal force f_int/ 
stiffness matrix Slocal from domain.state and then assemble to F_int/Ksparse
- 'globdat': GlobalData
- 'domain': Domain, finite element domain, for data structure
- 'Δt':  Float64, current time step size
Return F_int: Float64[neqs], internal force vector
Return Ksparse: Float64[neqs, neqns], sparse stiffness matrix

NNFEM.tfAssembleInternalForce — Method

Tensorflow version of assembling internal force vector, compute local internal 
force f_int and then assemble to F_int, which generates inverse problem automatically.

- 'domain': Domain, finite element domain, for data structure
- 'nn': Function strain -> stress, neural network constitutive law function 
- 'E_all': PyObject(Float64)[neles*nGauss, nstrains], neles*nGauss is the number of Gaussian quadrature points, 
           nstrain is the number of strain components. All strains for the current time-step
- 'DE_all': PyObject(Float64)[neles*nGauss, nstrains], neles*nGauss is the number of Gaussian quadrature points, 
            nstrain is the number of strain components. All strains for the previous time-step
- 'w∂E∂u_all': PyObject(Float64)[neles*nGauss, ndofs_per_element, nstrains], neles*nGauss is the number of Gaussian quadrature points, 
               ndofs_per_element is the number of freedoms per element, nstrain is the number of strain components.
               Multiplication of the Gaussian weight and ∂E∂u^T for current time-step, 
        
- 'σ0_all': PyObject(Float64)[neles*nGauss, nstrains], neles*nGauss is the number of Gaussian quadrature points, 
nstrain is the number of strain components.  All stresses for the last time-step

Return: internal force vector F_int, PyObject(Float64)[neqns] and the predicted stresses at the 
        current step σ_all, PyObject(Float64)[neles*nGauss, nstrains]

NNFEM.Domain — Type

Domain

Date structure for the computatational domain

nnodes: Int64, number of nodes (each quadratical quad element has 9 nodes)

nodes: Float64[nnodes, ndims], coordinate array of all nodes

neles: number of elements

elements: Element[neles], element array, each element is a struct

ndims: Int64, dimension of the problem space

state: Float64[nnodes*ndims] current displacement of all nodal freedoms, Float64[1:nnodes] are for the first direction.

Dstate: Float64[nnodes*ndims] previous displacement of all nodal freedoms, Float64[1:nnodes] are for the first direction.

'LM': Int64[neles][ndims], LM(e,d) is the global equation number(active freedom number) of element e's d th freedom, -1 means fixed (time-independent) Dirichlet -2 means time-dependent Dirichlet

'DOF': Int64[neles][ndims], DOF(e,d) is the global freedom number of element e's d th freedom

'ID': Int64[nnodes, ndims], ID(n,d) is the equation number(active freedom number) of node n's dth freedom, -1 means fixed (time-independent) Dirichlet -2 means time-dependent Dirichlet

'neqs': Int64, number of equations or active freedoms

'eqtodof': Int64[neqs], map from to equation number(active freedom number) to the freedom number (Int64[1:nnodes] are for the first direction)

'doftoeq': Bool[nnodes*ndims], map from freedom number(Int64[1:nnodes] are for the first direction) to booleans (active freedoms(equation number) are true)

'EBC': Int64[nnodes, ndims], EBC[n,d] is the displacement boundary condition of node n's dth freedom, -1 means fixed(time-independent) Dirichlet boundary nodes -2 means time-dependent Dirichlet boundary nodes

'g': Float64[nnodes, ndims], values for fixed(time-independent) Dirichlet boundary conditions of node n's dth freedom,

'FBC': Int64[nnodes, ndims], FBC[n,d] is the force load boundary condition of node n's dth freedom, -1 means constant(time-independent) force load boundary nodes -2 means time-dependent force load boundary nodes

'fext': Float64[neqs], constant(time-independent) force load boundary conditions for these freedoms

'time': Float64, current time

'npoints': Int64, number of points (each quadratical quad element has 4 points, npoints==nnodes, when porder==1)

'nodetopoint': Int64[nnodes]:map from node number to point point, -1 means the node is not a geometry point

'ii_stiff': Int64[], first index of the sparse matrix representation of the stiffness matrix

'jj_stiff': Int64[], second index of the sparse matrix representation of the stiffness matrix

'vvstiffele_indptr': Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the stiffness matrix

'vv_stiff': Float64[], values of the sparse matrix representation of the stiffness matrix

'iidfintdstress': Int64[], first index of the sparse matrix representation of the dfint_dstress matrix

'jjdfintdstress': Int64[], second index of the sparse matrix representation of the dfint_dstress matrix

'vvdfintdstresseleindptr': Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the dfint_dstress matrix

'vvdfintdstress': Float64[], values of the sparse matrix representation of the dfint_dstress matrix

'iidstraindstate': Int64[], first index of the sparse matrix representation of the dstrain_dstate matrix

'jjdstraindstate': Int64[], second index of the sparse matrix representation of the dstrain_dstate matrix

'vvdstraindstateeleindptr': Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the stiffness matrix

'vvdstraindstate': Float64[], values of the sparse matrix representation of the dstrain_dstate matrix

'history': Dict{String, Array{Array{Float64}}}, dictionary between string and its time-histories quantity Float64[ntime][]

NNFEM.Domain — Method

Creating a finite element domain

- `nodes`: Float64[nnodes, ndims], coordinate array of all nodes
- `elements`: Element[neles], element array, each element is a struct 
- `ndims`: Int64, dimension of the problem space 
- 'EBC':  Int64[nnodes, ndims], EBC[n,d] is the displacement boundary condition of node n's dth freedom,
       -1 means fixed(time-independent) Dirichlet boundary nodes
       -2 means time-dependent Dirichlet boundary nodes
- 'g':  Float64[nnodes, ndims], values for fixed(time-independent) Dirichlet boundary conditions of node n's dth freedom,
- 'FBC': Int64[nnodes, ndims], FBC[n,d] is the force load boundary condition of node n's dth freedom,
       -1 means constant(time-independent) force load boundary nodes
       -2 means time-dependent force load boundary nodes
- 'f':  Float64[nnodes, ndims], values for constant(time-independent) force load boundary conditions of node n's dth freedom,

Return: Domain, the finite element domain

NNFEM.GlobalData — Type

GlobalData

Store data for finite element update, assume the problem has n freedoms

state: Float64[n], displacement array at the current time, only for active freedoms, the ordering is based on the equation number, here ndims=2 is the dimension of the problem space

state: Float64[n], displacement array at the previous time

velo: Float64[n], velocity array at the current

acce: Float64[n], acceleration array at the current

time: float, current time

M: Float64[n,n] spares mass matrix

Mlumped: Float64[n] lumped mass array

MID: Float64[n, nd1] off-diagonal part of the mass matrix, between the active freedoms and the time-dependent Dirichlet freedoms, assume there are nd time-dependent Dirichlet freedoms

EBC_func: function Float64:t-> Float64[n_d1] float array, time-dependent Dirichlet boundary condition (ordering is direction first then node number, u1, u3, ... v1, v4 ...)

FBC_func: function Float64:t-> Float64[n_t1] float array, time-dependent load boundary condition (ordering is direction first then node number, u1, u3, ... v1, v4 ...)

NNFEM.setGeometryPoints! — Method

In the constructor 
Update the node_to_point map 

- `self`: Domain, finit element domain
- 'npoints': Int64, number of points (each quadratical quad element has 4 points, npoints==nnodes, when porder==1)
- 'node_to_point': Int64[nnodes]:map from node number to point point, -1 means the node is not a geometry point

NNFEM.setNeumannBoundary! — Method

In the constructor
Update the external force vector fext
- 'self': Domain
- 'FBC': Int64[nnodes, ndims], FBC[n,d] is the force load boundary condition of node n's dth freedom,
       -1 means constant(time-independent) force load boundary nodes
       -2 means time-dependent force load boundary nodes
- 'f':  Float64[nnodes, ndims], values for constant(time-independent) force load boundary conditions of node n's dth freedom,

NNFEM.updateDomainStateBoundary! — Method

Update time dependent boundary of state fext in Domain based on the time-dependent boundary functions in GlobalData.

'self': Domain

'globaldat': GlobalData

NNFEM.updateStates! — Method

At each time step
Update the state and Dstate in Domain from GlobalData
state and Dstate in GlobalData are only for active freedoms (equations)
state and Dstate in Domain are only for all freedoms, they are used for constructing the internal force and/or stiffness matrix
update the state and acc history of the Domain
- 'self': Domain
- 'globaldat': GlobalData

NNFEM.assembleSparseMatrixPattern! — Method

Compute constant stiff, dfint_dstress, dstrain_dstate matrix patterns
- 'self': Domain

NNFEM.commitHistory — Method

Update current step strain, stress in the history map of the Domain
strain and stress are both Float[ngp, nstrain], ngp is the number of Gaussian quadrature point
nstrain=1 for 1D and nstrain = 3 for 2D 
The strain is in Voigt notation

NNFEM.getCoords — Method

Get the coordinates of several nodes (possibly in one element)
- 'self': Domain
- 'el_nodes': Int64[n], node array

Return: Float64[n, ndims], the coordinates of these nodes

NNFEM.getDofs — Method

Get the global freedom numbers of the element
- 'self': Domain
- 'iele': Int64, element number

Return: Int64[], the global freedom numbers of the element (ordering in local element ordering)

NNFEM.getDstate — Method

Get the displacements of several nodes (possibly in one element) at the previous time step
- 'self': Domain
- 'el_nodes': Int64[n], node array

Return: Float64[n, ndims], the displacements of these nodes at the previous time step

NNFEM.getEqns — Method

Get the equation numbers(active freedom numbers) of the element
- 'self': Domain
- 'iele': Int64, element number

Return: Int64[], the equation numbers(active freedom numbers) of the element (ordering in local element ordering)

NNFEM.getExternalForce! — Method

Compute external force vector, including external force load and time-dependent Dirichlet boundary conditions
The function needs to be called after
function updateDomainStateBoundary!(self::Domain, globaldat::GlobalData)
which computes the external force vector from external force load

- 'self': Domain
- 'globaldat': GlobalData
- 'fext': Float64[neqs], container for the external force vector

return external force vector at the present step

NNFEM.getState — Method

Get the displacements of several nodes (possibly in one element)
- 'self': Domain
- 'el_nodes': Int64[n], node array

Return: Float64[n, ndims], the displacements of these nodes

NNFEM.setDirichletBoundary! — Method

In the constructor 
Update the fixed (time-independent Dirichlet boundary) state entries 
Build LM, and DOF array
- 'self': Domain
- 'EBC':  Int64[nnodes, ndims], EBC[n,d] is the displacement boundary condition of node n's dth freedom,
       -1 means fixed(time-independent) Dirichlet boundary nodes
       -2 means time-dependent Dirichlet boundary nodes
- 'g':  Float64[nnodes, ndims], values for fixed(time-independent) Dirichlet boundary conditions of node n's dth freedom,
''

Solvers

NNFEM.DynamicMatLawLoss — Method

- 'domain': Domain, finite element domain, for data structure
- 'E_all':  Float64[NT+1, neles*nGauss, nstrains], all strains for the whole simulation from training data
- 'w∂E∂u_all': Float64[NT+1, neles*nGauss, ndofs_per_element, nstrains], 
            multiplication of the Gaussian weight and ∂E∂u^T for the whole simulation from training data 
- 'F_tot': Float64[NT, neqs] approximated internal force for the whole simulation, from time n=1 to time n=NT

form the loss function, based on dynamic equilibrium 
    (Mddu + fint(NN, E, DE) + MIDddu_bc = fext

we have  
    F_tot  = fext - MIDddu_bc - Mddu
loss = ∑ ||fint(NN, E, DE) - F_tot||^2

todo
We assume the time step is constant for the training data

NNFEM.DynamicMatLawLoss — Method

- 'domain': Domain,  finite element domain
- 'globdat': GlobalData,  finite element data structure
- 'state_history': Float64[NT+1][ndofs], displace history of all time steps and all nodes, including time 0
                hcat(state_history...) gives a matrix of Float64[NT+1, ndofs]
- 'fext_history': Float64[NT+1][neqns] or Float64[NT][neqns]: external force load of all time steps and all equations, 
                with or without the initial step
                hcat(fext_history...) gives a matrix of size(neqns, NT+1)
nn: Neural network
Δt: time step size

compute loss function from state and external force history

NNFEM.LSfittingStress — Method

Fit the stress from the strain and external load conditions, 
For quadratic element, nx, ny, ... elements in each direction
The number of equations are neqs ≈ 2*(2nx + 1)*(2ny + 1)
Linear approach : Assume the stress in each element is constant, the number of unknows are 3*nx*ny
Constant approach: Assume the stress in each element is linear, the number of unknows are 3*(nx+1)*(ny+1) on each nodes
Least Square problem is solving for stress

- 'domain': Domain 
- 'globdat': GlobalData
- 'state_history': Float64[NT+1, ndofs], displacement fields for each freedoms
- 'F_ext': Float64[neqns, NT+1] or Float64[neqns, NT+1], external force vector for each equations
- 'Δt': Float64, time step size
- 'method': String "Constant" or "Linear" for Constant approach or Linear approach

Return: 
- 'E_all':  Float64[NT+1, neles*ngps_per_elem, nstrain], strain fields
- 'S_all':  Float64[NT+1, neles*ngps_per_elem, nstrain], fitted stress fields at each Gaussian quadratic point

Remark: Turss 1D requires method == "Constant", otherwise the system is under constraint

NNFEM.LSfittingStressHelper — Method

Compute the Fint, dFintdScomp, Eall, Sall and dfintdstress, based on the state in domain and the stress in Scomp

'domain': Domain

'S_comp': Float64[neles, nstrain] or Float64[npoints, nstrain], the stresses in each element(Constant) or at each geometric point(Linear)

'method': String "Constant" or "Linear" for Constant approach or Linear approach

Return Fint: Float64[neqns], constructed from state in domain dFintdScomp: Float64[neqs, size(Scomp,1)*nstrain], sparse matrix representation Eall: Float64[nelesngpsperelem, nstrain], strains at each Gaussian quadrature point ,constructed from state in domain S_all: Float64[nelesngpsperelem, nstrain], stresses at each Gaussian quadrature point ,constructed from S_comp

NNFEM.preprocessing — Method

Postprocess the data from domain.history and F_ext
- 'domain': Domain, domain.history["state"] is Float64[NT+1][ndofs] 
- 'globdat': GlobalData, for mass matrix
- 'F_ext': Float64[neqns, NT+1] or Float64[neqns, NT], external force history with/without the initial step
- 'Δt': Float64, time step size

Return:
- 'F_tot': Float64[NT, neqns], approxmiated internal force
- 'E_all': Float64[NT+1, neles*nGauss, nstrains], all strains for the whole simulation from training data
- 'w∂E∂u_all': Float64[NT+1, neles*nGauss, neqns_per_elem, nstrains], multiplication of the Gaussian weight and ∂E∂u^T 
for the whole simulation from training data 

todo
We assume the time step is constant for the training data

NNFEM.AdaptiveSolver — Method

Adaptive Solver, solve the whole process, if this step fails, redo the step with half of
the time step size, when there are continuing 5 successful steps, double the step size when dt < T/NT

- 'solvername': String,  so far only NewmarkSolver is supported
- 'globdat', GlobalData
- 'domain', Domain
- 'T', Float64, total simulation time
- 'NT', Int64, planned time steps
- 'args', Dict{String, Value}, arguments for the solver

return globdat, domain, ts, here ts is Float64[nteps+1]

todo For the first time step a0 = M^{-1}(- f^{int}(u0) + f^{ext}_0)

We assume globdat.acce[:] = a_0 and so far initialized to 0

NNFEM.ExplicitSolver — Method

Central Difference explicit solver

'Δt': Float64, time step size

'globdat', GlobalData

'domain', Domain

Central Difference explicit solver for M a + fint(u) = fext(u), with lumped mass matrix (MID = 0) a, v, u are acceleration, velocity and displacement

\[\begin{align} u_{n+1} =& u_n + dtv_n + dt^2/2 a_n \\ v_{n+1} =& v_n + dt/2(a_n + a_{n+1}) \\ M a_{n+1} + f^{int}(u_{n+1}) =& f^{ext}_{n+1} \\ M a_{n+1} =& f^{ext}_{n+1} - f^{int}(u_{n+1}) \\ \end{align}\]

use the current states a, v, u, time in globdat, and update these stetes to next time step update domain history

todo For the first time step a0 = M^{-1}(- f^{int}(u0) + f^{ext}_0)

We assume globdat.acce[:] = a_0 and so far initialized to 0

NNFEM.NewmarkSolver — Function

NewmarkSolver (Generalized-alpha) implicit solver

'Δt': Float64, time step size

'globdat', GlobalData

'domain', Domain

'αm', Float64

'αf', Float64

'ε', Float64, absolute error for Newton convergence

'ε0', Float64, relative error for Newton convergence

'maxiterstep', Int64, maximum iteration number for Newton convergence

'η', Float64, Newton step size at the first iteration

'failsafe', Bool, if failsafe is true, when the Newton fails to converge, revert back, and return false

Implicit solver for $Ma + fint(u) = fext$ $a$, $v$, $u$ are acceleration, velocity and displacement respectively.

\[u_{n+1} = u_n + dtv_n + dt^2/2 ((1 - 2\beta)a_n + 2\beta a_{n+1}) v_{n+1} = v_n + dt((1 - \gamma)a_n + \gamma a_{n+1}) 2\beta = 0.5*(1 - αm + αf)^2 \gamma = 0.5 - \alpha_m + \alpha_f \]

\[a_{n+1-\alpha_m} = (1-\alpha_m)a_{n+1} + \alpha_m a_{n} v_{n+1-\alpha_f} = (1-\alpha_f)v_{n+1} + \alpha_f v_{n} u_{n+1-\alpha_f} = (1-\alpha_f)u_{n+1} + \alpha_f u_{n} M a_{n+1-\alpha_m} + f^{int}(u_{n+1-\alpha_f}) = f^{ext}_{n+1-\alpha_f}\]

'a_{n+1}' is solved by

\[M ((1-\alpha_m)a_{n+1} + \alpha_m a_{n}) + f^{int}((1-\alpha_f)(u_n + dtv_n + dt^2/2 ((1 - 2\beta)a_n + 2\beta a_{n+1}))) + \alpha_f u_{n}) = f^{ext}_{n+1-\alpha_f}\]

As for '\alpham' and '\alphaf'

\[\alpha_m = (2\rho_{\infty} - 1)/(\rho_{\infty} + 1) \alpha_f = \rho_{\infty}/(\rho_{\infty} + 1)\]

use the current states a, v, u, time in globdat, and update these stetes to next time step update domain history, when failsafe is true, and Newton's solver fails, nothing will be changed.

todo For the first time step a0 = M^{-1}(- f^{int}(u0) + f^{ext}_0)

We assume globdat.acce[:] = a_0 and so far initialized to 0 We also assume the external force is conservative (it does not depend on the current deformation)

NNFEM.StaticSolver — Function

Static implicit solver

'globdat', GlobalData

'domain', Domain

'loaditerstep', Int64, load stepping steps

'ε', Float64, absolute error for Newton convergence

'maxiterstep', Int64, maximum iteration number for Newton convergence

Solver for $fint(u) = fext$ with load stepping, $u$ is the displacement. Iteratively solve u_{i}

\[f^{int}(u_i) = i/loaditerstep f^{ext}\]

todo We assume the external force is conservative (it does not depend on the current deformation)

Utilities

NNFEM.orthotropic_H — Method

orthotropic_H(y::PyObject)
orthotropic_H(o::Array)

Creates a symmetric matrix from 4 parameters

\[H = \begin{bmatrix} y_1 & y_2 & 0\\ y_2 & y_3 & 0 \\ 0 & 0 & y_4 \end{bmatrix}\]

NNFEM.spd_Chol_Orth — Method

spd_Chol_Orth(o::Array)
spd_Chol_Orth(o::PyObject)

Creates a SPD matrix from 4 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.spd_Cholesky — Method

spd_Cholesky(o::Array)
spd_Cholesky(o::PyObject)

Creates a SPD matrix from 6 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_4 & \\ o_3 & o_5 & o_6 \end{bmatrix}\]

NNFEM.spd_H — Method

spd_H(o::PyObject, H0::Array{Float64,2})
spd_H(o::Array, H0::Array)

Creates a SPD matrix from 3 scalars

\[H = H_0 - \frac{H_0nn'H_0}{1+n'H_0n}\]

NNFEM.spd_zero_to_H — Method

spd_zero_to_H(o::Array)
spd_zero_to_H(o::Array, H0inv::Array{Float64,2})

Creates a SPD matrix from 4 scalars.

\[A = (H_0^{-1} +LL')^{-1}\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.sym_H — Method

sym_H(y::PyObject)
sym_H(o::Array)

Creates a symmetric matrix from 6 parameters

\[H = \begin{bmatrix} y_1 & y_2 & y_3\\ y_2 & y_4 & y_5 \\ y_3 & y_5 & y_6 \end{bmatrix}\]

NNFEM.gradtest — Method

gradtest(f::Function, x0::Array{Float64}; scale::Float64 = 1.0)

Testing the gradients of a vector function f. Here x0 is $n$-dimensional vector, f takes an $n$-dimensional vector as inputs, and outputs a $m$ dimensional vector. The function tests the second order convergence of f

\[\|f(x_0+\gamma c) - f(x_0) - \gamma \nabla f(x_0) c\|_2 = \mathcal{O}(\gamma^2)\]