API Reference

Elements

NNFEM.FiniteStrainContinuum — Type

FiniteStrainContinuum(coords::Array{Float64}, elnodes::Array{Int64}, props::Dict{String, Any}, ngp::Int64=2)

NNFEM.FiniteStrainContinuum — Type

FiniteStrainContinuum

Constructs a finite strain element.

eledim: spatial dimension of the element (default = 2).

mat: constitutive law, a length #elem vector of materials such as PlaneStress

elnodes: the node indices in this finite element, an integer array

props: property dictionary

coords: coordinates of the vertices of the element

dhdx, weights, hs: data for integral

stress: stress at each quadrature points

Example

#   Local degrees of freedom 
#   4 ---- 3
#
#   1 ---- 2

nx = 10
ny = 5
h = 0.1
element = FiniteStrainContinuum[]
prop = Dict("name"=> "PlaneStrain", "rho"=> 0.0876584, "E"=>0.07180760098, "nu"=>0.4)
for j = 1:ny
    for i = 1:nx 
        n = (nx+1)*(j-1) + (i-1)+1
        elnodes = [n, n + 1, n + 1 + (nx + 1), n + (nx + 1)]
        ngp = 3 # 3 x 3 Gauss points per element 
        coords = [(i-1)*h (j-1)*h
                    i*h (j-1)*h
                    i*h j*h
                    (i-1)*h j*h]
        push!(element, FiniteStrainContinuum(coords,elnodes, prop, ngp))
    end
end

NNFEM.SmallStrainContinuum — Type

SmallStrainContinuum(coords::Array{Float64}, elnodes::Array{Int64}, props::Dict{String, Any}, ngp::Int64=2)

NNFEM.SmallStrainContinuum — Type

SmallStrainContinuum

Constructs a small strain element.

eledim: spatial dimension of the element (default = 2).

mat: constitutive law, a length #elem vector of materials such as PlaneStress

elnodes: the node indices in this finite element, an integer array

props: property dictionary

coords: coordinates of the vertices of the element

dhdx: list of ngp shape functions for first order derivatives $\nabla \phi(x)$ (ndof×2) on the Gaussian points

weights: weight vector of length n_gauss_points, for numerical quadrature

hs: list of ngp shape functions for function values $\phi(x)$ (length ndof vectors) on the Gaussian points

stress: stress at each quadrature points; this field is reserved for visualization.

Example

#   Local degrees of freedom 
#   4 ---- 3
#
#   1 ---- 2

nx = 10
ny = 5
h = 0.1
element = SmallStrainContinuum[]
prop = Dict("name"=> "PlaneStrain", "rho"=> 0.0876584, "E"=>0.07180760098, "nu"=>0.4)
for j = 1:ny
    for i = 1:nx 
        n = (nx+1)*(j-1) + (i-1)+1
        elnodes = [n, n + 1, n + 1 + (nx + 1), n + (nx + 1)]
        ngp = 3 # 3 x 3 Gauss points per element 
        coords = [(i-1)*h (j-1)*h
                    i*h (j-1)*h
                    i*h j*h
                    (i-1)*h j*h]
        push!(element, SmallStrainContinuum(coords,elnodes, prop, ngp))
    end
end

Materials

NNFEM.PlaneStress — Type

PlaneStress

Creates a plane stress element

H: Linear elasticity matrix, $3\times3$

E: Young's modulus

ν: Poisson's ratio

ρ: density

σ0: stress at the last time step

σ0_: (for internal use), stress to be updated in commitHistory

ε0: strain at the last time step

ε0_: (for internal use), strain to be updated in commitHistory

Example

prop = Dict("name"=> "PlaneStress", "rho"=> 0.0876584, "E"=>0.07180760098, "nu"=>0.4)
mat = PlaneStress(prop)

NNFEM.PlaneStress — Method

PlaneStress(prop::Dict{String, Any})

prop should contain at least the following three fields: E, nu, rho

NNFEM.PlaneStrain — Type

PlaneStrain

Creates a plane strain element

H: Linear elasticity matrix, $3\times3$

E: Young's modulus

ν: Poisson's ratio

ρ: density

σ0: stress at the last time step

σ0_: (for internal use), stress to be updated in commitHistory

ε0: strain at the last time step

ε0_: (for internal use), strain to be updated in commitHistory

Example

prop = Dict("name"=> "PlaneStrain", "rho"=> 0.0876584, "E"=>0.07180760098, "nu"=>0.4)
mat = PlaneStrain(prop)

NNFEM.PlaneStrain — Method

PlaneStrain(prop::Dict{String, Any})

prop should contain at least the following three fields: E, nu, rho

NNFEM.PlaneStressIncompressibleRivlinSaunders — Type

Pascon, João Paulo. "Large deformation analysis of plane-stress hyperelastic problems via triangular membrane finite elements." International Journal of Advanced Structural Engineering (2019): 1-20.

Assembly

NNFEM.Domain — Type

Domain

Date structure for the computatational domain.

nnodes: Int64, number of nodes (each quadratical quad element has 9 nodes)

nodes: Float64[nnodes, ndims], coordinate array of all nodes

neles: number of elements

elements: a list of neles element arrays, each element is a struct

ndims: Int64, dimension of the problem space

state: a matrix of size nnodes×ndims. Current displacement of all nodal freedoms, state[1:nnodes] are for the first direction.

Dstate: nnodes×ndims. Previous displacement of all nodal freedoms, Dstate[1:nnodes] are for the first direction.

LM: Int64[neles][ndims], LM(e,d) is the global equation number (active freedom number) of element e's d th freedom, -1 means fixed (time-independent) Dirichlet -2 means time-dependent Dirichlet

DOF: a matrix of size neles×ndims, DOF(e,d) is the global freedom number of element e's d th freedom

ID: a matrix of size nnodes×ndims. ID(n,d) is the equation number (active freedom number) of node n's $d$-th freedom, -1 means fixed (time-independent) Dirichlet -2 means time-dependent Dirichlet

neqs: Int64, number of equations, a.k.a., active freedoms

eq_to_dof: an integer vector of length neqs, map from to equation number (active freedom number) to the freedom number (Int64[1:nnodes] are for the first direction)

dof_to_eq: a bolean array of size nnodes×ndims, map from freedom number(Int64[1:nnodes] are for the first direction) to booleans (active freedoms(equation number) are true)

EBC: Int64[nnodes, ndims], EBC[n,d] is the displacement boundary condition of node n's dth freedom, -1 means fixed(time-independent) Dirichlet boundary nodes -2 means time-dependent Dirichlet boundary nodes

g: Float64[nnodes, ndims], values for fixed(time-independent) Dirichlet boundary conditions of node n's dth freedom,

FBC: Int64[nnodes, ndims], FBC[n,d] is the force load boundary condition of node n's dth freedom, -1 means constant(time-independent) force load boundary nodes -2 means time-dependent force load boundary nodes

fext: Float64[neqs], constant (time-independent) force load boundary conditions for these freedoms

time: Float64, current time

npoints: Int64, number of points (each quadratical quad element has 4 points, npoints==nnodes, when porder==1)

node_to_point: Int64[nnodes]:map from node number to point point, -1 means the node is not a geometry point

Auxilliry Data Structures

ii_stiff: Int64[], first index of the sparse matrix representation of the stiffness matrix

jj_stiff: Int64[], second index of the sparse matrix representation of the stiffness matrix

vv_stiff_ele_indptr: Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the stiffness matrix

vv_stiff: Float64[], values of the sparse matrix representation of the stiffness matrix

ii_dfint_dstress: Int64[], first index of the sparse matrix representation of the dfint_dstress matrix

jj_dfint_dstress: Int64[], second index of the sparse matrix representation of the dfint_dstress matrix

vv_dfint_dstress_ele_indptr: Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the dfint_dstress matrix

vv_dfint_dstress: Float64[], values of the sparse matrix representation of the dfint_dstress matrix

ii_dstrain_dstate: Int64[], first index of the sparse matrix representation of the dstrain_dstate matrix

jj_dstrain_dstate: Int64[], second index of the sparse matrix representation of the dstrain_dstate matrix

vv_dstrain_dstate_ele_indptr: Int64[], Int64[e] is the first index entry for the e's element of the sparse matrix representation of the stiffness matrix

vv_dstrain_dstate: Float64[], values of the sparse matrix representation of the dstrain_dstate matrix

history: Dict{String, Array{Array{Float64}}}, dictionary between string and its time-histories quantity Float64[ntime][]

NNFEM.GlobalData — Type

GlobalData

Store data for finite element updates. Assume the problem has n freedoms,

state: a vector of length $n$. Displacement array at the current time, only for active freedoms. The ordering is based on the equation number.

Dstate: a vector of length $n$. Displacement array at the previous time.

velo: a vector of length $n$. Velocity array at the current time.

acce: a vector of length $n$. Acceleration array at the current time.

time: float, current time.

M: a matrix of size $n\times n$ spares mass matrix

Mlumped: a vector of length $n$ lumped mass array

MID: Float64[n, nd1] off-diagonal part of the mass matrix, between the active freedoms and the time-dependent Dirichlet freedoms, assume there are nd time-dependent Dirichlet freedoms

EBC_func: displacement $d$, velocity $v$, and acceleration $a$ from time-dependent Dirichlet boundary conditions

\[d, v, a = \text{EBC\_func}(time)\]

The length of each output is the same as number of "-2" in EBC array. The ordering is direction major, i.e., $u_1, u_3, \ldots, v_1, v_3, \ldots$

FBC_func: time-dependent load boundary condition.

\[f = \text{FBC\_func}(time)\]

Here $f$ is a vector. Its length is the same as number of "-2" in FBC array. The ordering is direction major, i.e., $u_1, u_3, \ldots, v_1, v_3, \ldots$

NNFEM.commitHistory — Function

commitHistory(domain::Domain)

Update current step strain and stress in the history map of the domain. This is essential for visualization and time dependent constitutive relations.

NNFEM.setDirichletBoundary! — Function

setDirichletBoundary!(self::Domain, EBC::Array{Int64}, g::Array{Float64})

Bookkeepings for Dirichlet boundary conditions. Only called once in the constructor of domain. It updates the fixed (time-independent Dirichlet boundary) state entries and builds both LM and DOF arrays.

self: Domain

EBC: Int64[nnodes, ndims], EBC[n,d] is the displacement boundary condition of node n's dth freedom,

∘ -1 means fixed(time-independent) Dirichlet boundary nodes

∘ -2 means time-dependent Dirichlet boundary nodes

g: Float64[nnodes, ndims], values for fixed (time-independent) Dirichlet boundary conditions of node n's dth freedom,

NNFEM.setNeumannBoundary! — Function

Bookkeepings for Dirichlet boundary conditions. Only called once in the constructor of domain. It updates the fixed (time-independent Neumann boundary) state entries and builds both LM and DOF arrays.

self: Domain

FBC: Int64[nnodes, ndims], FBC[n,d] is the displacement boundary condition of node n's dth freedom,

∘ -1 means fixed (time-independent) Neumann boundary nodes

∘ -2 means time-dependent Dirichlet boundary nodes

f: Float64[nnodes, ndims], values for fixed (time-independent) Neumann boundary conditions of node n's dth freedom,

NNFEM.updateStates! — Function

updateStates!(self::Domain, globaldat::GlobalData)

At each time step, updateStates! needs to be called to update the full state and Dstate in domain from active ones in globaldat.

NNFEM.updateDomainStateBoundary! — Function

updateDomainStateBoundary!(self::Domain, globaldat::GlobalData)

If there exists time-dependent boundary conditions, updateDomainStateBoundary! must be called to update the boundaries in domain. This function is called by updateStates!

NNFEM.getExternalForce! — Function

getExternalForce!(self::Domain, globaldat::GlobalData, fext::Union{Missing,Array{Float64}}=missing)

Computes external force vector, including external force load and time-dependent Dirichlet boundary conditions.

Info

The function needs to be called after updateDomainStateBoundary!, which computes the external force vector from external force load

Solvers

Missing docstring.

Missing docstring for ExplicitSolver. Check Documenter's build log for details.

NNFEM.SolverInitial! — Function

SolverInitial!(Δt::Int64, globdat::GlobalData, domain::Domain)

You need to call SolverInitial! before the first time step, if $f^{ext}_0 \neq 0$

\[a_0 = M^{-1}(- f^{int}(u_0) + f^{ext}_0)\]

Utilities

NNFEM.readMesh — Method

readMesh(gmshFile::String)

Reads a gmsh file and extracts element, coordinates and boundaries.

NNFEM.orthotropic_H — Method

orthotropic_H(y::PyObject)
orthotropic_H(o::Array)

Creates a symmetric matrix from 4 parameters

\[H = \begin{bmatrix} y_1 & y_2 & 0\\ y_2 & y_3 & 0 \\ 0 & 0 & y_4 \end{bmatrix}\]

NNFEM.spd_Chol_Orth — Method

spd_Chol_Orth(o::Array)
spd_Chol_Orth(o::PyObject)

Creates a SPD matrix from 4 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.spd_Cholesky — Method

spd_Cholesky(o::Array)
spd_Cholesky(o::PyObject)

Creates a SPD matrix from 6 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_4 & \\ o_3 & o_5 & o_6 \end{bmatrix}\]

NNFEM.spd_H — Method

spd_H(o::PyObject, H0::Array{Float64,2})
spd_H(o::Array, H0::Array)

Creates a SPD matrix from 3 scalars

\[H = H_0 - \frac{H_0nn'H_0}{1+n'H_0n}\]

NNFEM.spd_zero_to_H — Method

spd_zero_to_H(o::Array)
spd_zero_to_H(o::Array, H0inv::Array{Float64,2})

Creates a SPD matrix from 4 scalars.

\[A = (H_0^{-1} +LL')^{-1}\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.sym_H — Method

sym_H(y::PyObject)
sym_H(o::Array)

Creates a symmetric matrix from 6 parameters

\[H = \begin{bmatrix} y_1 & y_2 & y_3\\ y_2 & y_4 & y_5 \\ y_3 & y_5 & y_6 \end{bmatrix}\]

NNFEM.gradtest — Method

gradtest(f::Function, x0::Array{Float64}; scale::Float64 = 1.0)

Testing the gradients of a vector function f. Here x0 is $n$-dimensional vector, f takes an $n$-dimensional vector as inputs, and outputs a $m$ dimensional vector. The function tests the second order convergence of f

\[\|f(x_0+\gamma c) - f(x_0) - \gamma \nabla f(x_0) c\|_2 = \mathcal{O}(\gamma^2)\]

Internals