API Reference

Elements

NNFEM.FiniteStrainContinuum — Type

FiniteStrainContinuum

Implements the finite strain element.

eledim: spatial dimension of the element (default = 2).

mat: constitutive law, a length #elem vector of materials such as PlaneStress

elnodes: the node indices in this finite element, an integer array

props: property dictionary

coords: coordinates of the vertices of the element

dhdx, weights, hs: data for integral

stress: stress at each quadrature points

Materials

NNFEM.PlaneStrain — Type

NNFEM.PlaneStressIncompressibleRivlinSaunders — Type

Pascon, João Paulo. "Large deformation analysis of plane-stress hyperelastic problems via triangular membrane finite elements." International Journal of Advanced Structural Engineering (2019): 1-20.

Assembly

NNFEM.assembleMassMatrix! — Method

compute constant mass matrix
due to the time-dependent Dirichlet boundary condition
mass matrix = M,    MID
              MID'  MDD
save M and MID and lump(M)

NNFEM.tfAssembleInternalForce — Method

domain   : finite element domain, for data structure
nn: Neural network constitutive law, with 
E_all    : all strains for the current time-step, with size (neles*nGauss, nstrains)
DE_all   : all strains for the last time-step, with size (neles*nGauss, nstrains)
w∂E∂u_all: multiplication of the Gaussian weight and ∂E∂u^T for current time-step, 
           with size (neles*nGauss, ndofs_per_element, nstrains)
σ0_all   : all stresses for the last time-step, with size (neles*nGauss, nstrains)

compute local internal force f_int and then assemble to F_int

NNFEM.Domain — Method

Creating a finite element domain
nodes: n by 2 float 64, node coordinates
elements: element list
ndims: 2
EBC: n by 2 Int64, nodal Dirichlet boundary condition, -1 time-independent, -2 time-dependent
g: n by 2 Float64, values for nodal time-independent Dirichlet boundary condition
FBC: n by 2 Int64, nodal force boundary condition, -1 time-independent, -2 time-dependent
f: n by 2 Float64, values for nodal force time independent force boundary condition

NNFEM.convertState — Method

:param self: Domain
:param state : 1D array to convert
:param compress_or_expand  : "Compress" or "Expand" 

"Compress", the state has all freedoms on all nodes, remove these freedoms on EBC
"Expand",   the state has only active freedoms on active nodes (active means not prescribed), 
            set these freedoms on EBC to 0

:return:

NNFEM.setGeometryPoints! — Method

:param npoints, the mesh contains npoints
:node_to_point, a map from node to point, -1 means the node is not the geometry point
:return:

NNFEM.setNeumannBoundary! — Method

:param FBC[n, d] is the boundary condition of node n's dth freedom,
    -1 means fixed force load boundary condition
    -2 means time dependent force load boundary condition
:param f[n, d] is the fixed force load value
:return:

NNFEM.updateDomainStateBoundary! — Method

Update domain time dependent boundary information.

NNFEM.updateStates! — Method

:param disp: neqs array
:param vel : neqs array

update Dstate in Domain, update state in Domain
:return:

NNFEM.assembleSparseMatrixPattern! — Method

compute constant stiff matrix pattern

NNFEM.getDofs — Method

:param el_nodes: 1d array
:return: the corresponding dofs ids, u0,u1, .., v0, v1, ..

NNFEM.getExternalForce! — Method

the external force include the Dirichlet boundary condition effect

NNFEM.setDirichletBoundary! — Method

:param EBC[n, d] is the boundary condition of of node n's dth freedom,
    -1 means fixed Dirichlet boundary nodes
    -2 means time dependent Dirichlet boundary nodes
:param g[n, d] is the fixed Dirichlet boundary value

:param nbc:
:return:

Solvers

NNFEM.DynamicMatLawLoss — Method

domain   : finite element domain, for data structure
E_all    : all strains for the whole simulation, with size (NT+1, neles*nGauss, nstrains)
w∂E∂u_all: multiplication of the Gaussian weight and ∂E∂u^T for the whole simulation, 
           with size (NT+1, neles*nGauss, ndofs_per_element, nstrains)
F_tot : approximated internal force for the whole simulation, with size(NT, ndofs), 
        from time n=1 to time n=NT

form the loss function, based on dynamic equilibrium 
    (Mddu + fint(NN, E, DE) + MIDddu_bc = fext

loss = ∑ ||fint(NN, E, DE) - (fext - MIDddu_bc - Mddu)||^2

NNFEM.DynamicMatLawLoss — Method

domain   : finite element domain
globdat  : finite element data structure
state_history : displace history of all time steps and all nodes, 
                a list of NT+1  ndof-displacement vectors, including time 0
                hcat(state_history...) gives a matrix of size(ndof-displacement, NT+1)
fext_history  : external force load of all time steps and all nodes, 
                a list of NT+1 ndof-external-force vectors, including time 0
                hcat(fext_history...) gives a matrix of size(ndof-external-force, NT+1)
nn: Neural network
Δt: time step size

compute loss function from state and external force history

NNFEM.LSfittingStress — Method

LSfittingStress(domain::Domain, globdat::GlobalData, state_history::Array{Float64}, F_ext::Array{Float64},Δt::Float64, method::String)

For quadratic element, nx, ny, ... elements in each direction The number of equations are neqs ≈ 2(2nx + 1)(2ny + 1) Plan 1: Assume the stress in each element is constant, the number of unknows are 3nxny Plan 2: Assume the stress in each element is linear, the number of unknows are 3(nx+1)(ny+1) on each nodes

NNFEM.LSfittingStressHelper — Method

Compute the stiff and dfintdstress, based on the state in domain and dstraindstate

NNFEM.preprocessing — Method

preprocessing(domain::Domain, globdat::GlobalData, F_ext::Array{Float64},Δt::Float64, n::Int64)

Same as preprocessing, except that only the first n steps are considered

NNFEM.preprocessing — Method

compute F_tot ≈ F_int , ane E_all

NNFEM.AdaptiveSolver — Method

Adaptive Solver, adjust the time step, if this step fails, redo the step with half of
the time step size

return globdat, domain

NNFEM.ExplicitSolver — Method

ExplicitSolver(Δt, globdat, domain)

Central Difference Explicit solver for Ma + fint = fext, a, v, u are acceleration, velocity and displacement

\[\begin{align} u_{n+1} =& u_n + dtv_n + dt^2/2 a_n \\ v_{n+1} =& v_n + dt/2(a_n + a_{n+1}) \\ M a_{n+1} + C v_{n+1} + R(u_{n+1}) =& P_{n+1} \\ (M + dt/2 C) a_{n+1} =& P_{n+1} - R(u_{n+1}) - C(v_n + dt/2 a_{n}) \\ \end{align}\]

Alternative, todo:
M a_n + C v_n + R(u_n) = P_n
M(u_{n+1} - 2u_n + u_{n-1}) + dt/2*C(u_{n+1} - u_{n-1}) + dt^2 R(u_n) = dt^2 P_n
(M + dt/2 C) u_{n+1} = dt^2(P_n - R(u_n) + dt/2 C u_{n-1} + M(2u_n - u_{n-1})

For the first step
u_{-1} = u_0 - dt*v_0 + dt^2/2 a_0
a_0 = M^{-1}(-Cv_0 - R(u_0) + P_0)

NNFEM.NewmarkSolver — Function

NewmarkSolver(Δt, globdat, domain, αm = -1.0, αf = 0.0, ε = 1e-8, ε0 = 1e-8, maxiterstep=100, η = 1.0, failsafe = false)

Implicit solver for $Ma + C v + R(u) = P$ $a$, $v$, $u$ are acceleration, velocity and displacement respectively.

\[u_{n+1} = u_n + dtv_n + dt^2/2 ((1 - 2\beta)a_n + 2\beta a_{n+1}) v_{n+1} = v_n + dt((1 - gamma)a_n + gamma a_{n+1})\]

\[M a_{n+0.5} + f_{\mathrm{int}}(u_{n+0.f}) = fext_{n+0.5}\]

αm = (2\rho_oo - 1)/(\rho_oo + 1)
αf = \rho_oo/(\rho_oo + 1)

β2 = 0.5*(1 - αm + αf)^2
γ = 0.5 - αm + αf

absolution error ε = 1e-8, 
relative error ε0 = 1e-8  

return true or false indicating converging or not

NNFEM.StaticSolver — Function

StaticSolver(globdat, domain, loaditerstep = 10, ε = 1.e-8, maxiterstep=100)

Implicit solver for

\[f_{\mathrm{int}}(u) = f_{\mathrm{ext}}\]

$u$ is the displacement. We apply the Newton-Raphson algorithm

\[u_{n+1} = u_n - \nabla f_{\mathrm{int}}(u^n)^{-1} *( f_{\mathrm{int}}(u^n) - f_{\mathrm{ext}})\]

Utilities

NNFEM.orthotropic_H — Method

orthotropic_H(y::PyObject)
orthotropic_H(o::Array)

Creates a symmetric matrix from 4 parameters

\[H = \begin{bmatrix} y_1 & y_2 & 0\\ y_2 & y_3 & 0 \\ 0 & 0 & y_4 \end{bmatrix}\]

NNFEM.spd_Chol_Orth — Method

spd_Chol_Orth(o::Array)
spd_Chol_Orth(o::PyObject)

Creates a SPD matrix from 4 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.spd_Cholesky — Method

spd_Cholesky(o::Array)
spd_Cholesky(o::PyObject)

Creates a SPD matrix from 6 scalars.

\[A = LL'\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_4 & \\ o_3 & o_5 & o_6 \end{bmatrix}\]

NNFEM.spd_H — Method

spd_H(o::PyObject, H0::Array{Float64,2})
spd_H(o::Array, H0::Array)

Creates a SPD matrix from 3 scalars

\[H = H_0 - \frac{H_0nn'H_0}{1+n'H_0n}\]

NNFEM.spd_zero_to_H — Method

spd_zero_to_H(o::Array)
spd_zero_to_H(o::Array, H0inv::Array{Float64,2})

Creates a SPD matrix from 4 scalars.

\[A = (H_0^{-1} +LL')^{-1}\]

where

\[L = \begin{bmatrix} o_1 & & \\ o_2 & o_3 & \\ & & o_4 \end{bmatrix}\]

NNFEM.sym_H — Method

sym_H(y::PyObject)
sym_H(o::Array)

Creates a symmetric matrix from 6 parameters

\[H = \begin{bmatrix} y_1 & y_2 & y_3\\ y_2 & y_4 & y_5 \\ y_3 & y_5 & y_6 \end{bmatrix}\]

NNFEM.gradtest — Method

gradtest(f::Function, x0::Array{Float64}; scale::Float64 = 1.0)

Testing the gradients of a vector function f. Here x0 is $n$-dimensional vector, f takes an $n$-dimensional vector as inputs, and outputs a $m$ dimensional vector. The function tests the second order convergence of f

\[\|f(x_0+\gamma c) - f(x_0) - \gamma \nabla f(x_0) c\|_2 = \mathcal{O}(\gamma^2)\]